refactor loader to resourceprovider (site and stage)
