+++ core/src/test/java/org/dotwebstack/framework/site/siteResourceProviderTest.java
@@ -1,16 +1,15 @@
-package org.dotwebstack.framework.stage;
+package org.dotwebstack.framework.site;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import org.dotwebstack.framework.config.ConfigurationBackend;
 import org.dotwebstack.framework.config.ConfigurationException;
-import org.dotwebstack.framework.site.Site;
-import org.dotwebstack.framework.site.SiteLoader;
 import org.dotwebstack.framework.test.DBEERPEDIA;
 import org.dotwebstack.framework.vocabulary.ELMO;
 import org.eclipse.rdf4j.model.ValueFactory;
 import org.eclipse.rdf4j.model.impl.SimpleValueFactory;
 import org.eclipse.rdf4j.model.vocabulary.RDF;
+import org.eclipse.rdf4j.model.vocabulary.RDFS;
 import org.eclipse.rdf4j.query.GraphQuery;
 import org.eclipse.rdf4j.query.impl.IteratingGraphQueryResult;
 import org.eclipse.rdf4j.repository.sail.SailRepository;
@@ -25,21 +24,16 @@ import org.mockito.junit.MockitoJUnitRunner;
 
 import static org.hamcrest.CoreMatchers.*;
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.collection.IsCollectionWithSize.hasSize;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.when;
 
 @RunWith(MockitoJUnitRunner.class)
-public class StageLoaderTest {
+public class siteResourceProviderTest {
 
   @Rule
   public final ExpectedException thrown = ExpectedException.none();
-
-  @Mock
-  private SiteLoader siteLoader;
-
-  @Mock
-  private Site site;
 
   @Mock
   private ConfigurationBackend configurationBackend;
@@ -55,70 +49,62 @@ public class StageLoaderTest {
 
   private ValueFactory valueFactory = SimpleValueFactory.getInstance();
 
-  private StageLoader stageLoader;
+  private SiteResourceProvider siteResourceProvider;
 
   @Before
   public void setUp() {
-    stageLoader = new StageLoader(siteLoader, configurationBackend);
+    siteResourceProvider = new SiteResourceProvider(configurationBackend);
 
     when(configurationBackend.getRepository()).thenReturn(configurationRepository);
     when(configurationRepository.getConnection()).thenReturn(configurationRepositoryConnection);
     when(configurationRepositoryConnection.prepareGraphQuery(anyString())).thenReturn(graphQuery);
-
-    when(siteLoader.getSite(any())).thenReturn(site);
   }
 
   @Test
-  public void loadStage() {
+  public void loadSite() {
     // Arrange
     when(graphQuery.evaluate()).thenReturn(new IteratingGraphQueryResult(ImmutableMap.of(),
         ImmutableList.of(
-            valueFactory.createStatement(DBEERPEDIA.STAGE, RDF.TYPE, ELMO.STAGE),
-            valueFactory.createStatement(DBEERPEDIA.STAGE, ELMO.SITE_PROP, DBEERPEDIA.SITE),
-            valueFactory.createStatement(DBEERPEDIA.STAGE, ELMO.BASE_PATH_PROP, DBEERPEDIA.BASE_PATH))));
+            valueFactory.createStatement(DBEERPEDIA.SITE, RDF.TYPE, ELMO.SITE),
+            valueFactory.createStatement(DBEERPEDIA.SITE, ELMO.DOMAIN_PROP, DBEERPEDIA.DOMAIN))));
 
     // Act
-    stageLoader.load();
+    siteResourceProvider.loadResources();
 
     // Assert
-    assertThat(stageLoader.getNumberOfStages(), equalTo(1));
-    Stage stage = stageLoader.getStage(DBEERPEDIA.STAGE);
-    assertThat(stage, is(not(nullValue())));
-    assertThat(stage.getSite(), equalTo(site));
-    assertThat(stage.getBasePath(), equalTo(DBEERPEDIA.BASE_PATH.stringValue()));
+    assertThat(siteResourceProvider.getAll().entrySet(), hasSize(1));
+    assertThat(siteResourceProvider.get(DBEERPEDIA.SITE), is(not(nullValue())));
   }
 
   @Test
-  public void loadMultipleStages() {
+  public void loadsSeveralSites() {
     // Arrange
     when(graphQuery.evaluate()).thenReturn(new IteratingGraphQueryResult(ImmutableMap.of(),
         ImmutableList.of(
-            valueFactory.createStatement(DBEERPEDIA.STAGE, RDF.TYPE, ELMO.STAGE),
-            valueFactory.createStatement(DBEERPEDIA.STAGE, ELMO.SITE_PROP, DBEERPEDIA.SITE),
-            valueFactory.createStatement(DBEERPEDIA.STAGE, ELMO.BASE_PATH_PROP, DBEERPEDIA.BASE_PATH),
-            valueFactory.createStatement(DBEERPEDIA.SECOND_STAGE, RDF.TYPE, ELMO.STAGE),
-            valueFactory.createStatement(DBEERPEDIA.SECOND_STAGE, ELMO.SITE_PROP, DBEERPEDIA.SITE))));
+            valueFactory.createStatement(DBEERPEDIA.SITE, RDF.TYPE, ELMO.SITE),
+            valueFactory.createStatement(DBEERPEDIA.SITE, ELMO.DOMAIN_PROP, DBEERPEDIA.DOMAIN),
+            valueFactory.createStatement(DBEERPEDIA.SITE_NL, RDF.TYPE, ELMO.SITE),
+            valueFactory.createStatement(DBEERPEDIA.SITE_NL, ELMO.DOMAIN_PROP, DBEERPEDIA.DOMAIN_NL))));
 
     // Act
-    stageLoader.load();
+    siteResourceProvider.loadResources();
 
     // Assert
-    assertThat(stageLoader.getNumberOfStages(), equalTo(2));
+    assertThat(siteResourceProvider.getAll().entrySet(), hasSize(2));
   }
 
   @Test
-  public void expectsSite() {
+  public void loadCatchAllSite() {
     // Arrange
     when(graphQuery.evaluate()).thenReturn(new IteratingGraphQueryResult(ImmutableMap.of(),
         ImmutableList.of(
-            valueFactory.createStatement(DBEERPEDIA.STAGE, RDF.TYPE, ELMO.STAGE))));
-
-    // Assert
-    thrown.expect(ConfigurationException.class);
-    thrown.expectMessage(String.format("No <%s> site has been found for stage <%s>.", ELMO.SITE_PROP, DBEERPEDIA.STAGE));
+            valueFactory.createStatement(DBEERPEDIA.SITE, RDF.TYPE, ELMO.SITE))));
 
     // Act
-    stageLoader.load();
+    siteResourceProvider.loadResources();
 
+    // Assert
+    assertThat(siteResourceProvider.getAll().entrySet(), hasSize(1));
+    assertThat(siteResourceProvider.get(DBEERPEDIA.SITE), is(not(nullValue())));
   }
 }
